# Conflicts:
#	.idea/workspace.xml

+++ src/plugin/index-basic/src/java/org/apache/nutch/indexer/basic/BasicIndexingFilter.java
@@ -105,7 +105,7 @@ public class BasicIndexingFilter implements IndexingFilter {
     double _char = textContent.length();
     double _a = page.getOutlinks().size();
     if (_a == 0) {
-      _a = (double)page.getVariable("outlinks_count");
+      _a = (int)page.getVariable("outlinks_count");
     }
 
     if (textContent.isEmpty()) {
